# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Product API",
        "description": "API для управления товарами, арендой и покупками в системе Laravel Product",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
...
        "/api/rent": {
            "post": {
                "tags": [
                    "Rent"
                ],
                "summary": "Арендовать товар",
                "description": "Позволяет пользователю арендовать товар на определенное количество часов. Требуется авторизация.",
                "operationId": "rentProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Товар успешно арендован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OwnershipResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации или бизнес-логики",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар или пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/rent/extend": {
            "post": {
                "tags": [
                    "Rent"
                ],
                "summary": "Продлить аренду товара",
                "description": "Позволяет пользователю продлить существующую аренду товара на дополнительные часы. Требуется авторизация.",
                "operationId": "extendRent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExtendRentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Аренда успешно продлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OwnershipResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации или бизнес-логики (аренда истекла, недостаточно средств)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Аренда, товар или пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/info": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Получить информацию о пользователе",
                "description": "Возвращает полную информацию о текущем авторизованном пользователе, включая его владения товарами. Требуется авторизация.",
                "operationId": "getUserInfo",
                "responses": {
                    "200": {
                        "description": "Информация о пользователе получена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserInfoResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/balance": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Получить баланс пользователя",
                "description": "Возвращает текущий баланс авторизованного пользователя. Требуется авторизация.",
                "operationId": "getUserBalance",
                "responses": {
                    "200": {
                        "description": "Баланс пользователя получен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserBalanceResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Обновить баланс пользователя",
                "description": "Обновляет баланс пользователя. Требуется авторизация и специальные права доступа.",
                "operationId": "updateUserBalance",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBalanceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Баланс пользователя обновлен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserBalanceResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав для выполнения операции",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "title": "Ответ с ошибкой",
                "description": "Стандартный ответ API при ошибке",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string",
                        "example": "Товар не найден"
                    },
                    "errors": {
                        "description": "Детали ошибок (опционально)",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "Ответ с ошибками валидации",
                "description": "Ответ API при ошибках валидации",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string",
                        "example": "Ошибка валидации данных"
                    },
                    "errors": {
                        "description": "Детали ошибок валидации",
                        "properties": {
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Имя обязательно"
                                ]
                            },
                            "purchase_price": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Цена покупки должна быть числом"
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "RentRequest": {
                "title": "Запрос на аренду товара",
                "required": [
                    "product_id",
                    "unique_code",
                    "hours"
                ],
                "properties": {
                    "product_id": {
                        "description": "Идентификатор товара для аренды",
                        "type": "integer",
                        "example": 1
                    },
                    "unique_code": {
                        "description": "Уникальный код товара",
                        "type": "string",
                        "example": "PROD-001"
                    },
                    "hours": {
                        "description": "Количество часов аренды",
                        "type": "integer",
                        "minimum": 1,
                        "example": 24
                    }
                },
                "type": "object"
            },
            "ExtendRentRequest": {
                "title": "Запрос на продление аренды",
                "required": [
                    "ownership_id",
                    "additional_hours"
                ],
                "properties": {
                    "ownership_id": {
                        "description": "Идентификатор аренды",
                        "type": "integer",
                        "example": 1
                    },
                    "additional_hours": {
                        "description": "Дополнительные часы аренды",
                        "type": "integer",
                        "minimum": 1,
                        "example": 12
                    }
                },
                "type": "object"
            },
            "Ownership": {
                "title": "Владение товаром",
                "properties": {
                    "id": {
                        "description": "Идентификатор владения",
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "description": "Идентификатор товара",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "unique_code": {
                        "description": "Уникальный код товара",
                        "type": "string",
                        "example": "PROD-001"
                    },
                    "type": {
                        "description": "Тип владения",
                        "type": "string",
                        "enum": [
                            "rent",
                            "purchase"
                        ],
                        "example": "rent"
                    },
                    "expires_at": {
                        "description": "Дата окончания аренды (только для аренды)",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T12:00:00Z"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Дата последнего обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "OwnershipResponse": {
                "title": "Ответ с данными владения",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Товар успешно арендован"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Ownership"
                    }
                },
                "type": "object"
            },
            "UserWithBalance": {
                "title": "Пользователь с балансом",
                "properties": {
                    "id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "description": "Email пользователя",
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "balance": {
                        "description": "Баланс пользователя",
                        "type": "number",
                        "format": "integer",
                        "example": 1500.5
                    },
                    "created_at": {
                        "description": "Дата регистрации",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "ownerships": {
                        "description": "Владения пользователя",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Ownership"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateBalanceRequest": {
                "title": "Запрос на обновление баланса",
                "required": [
                    "amount"
                ],
                "properties": {
                    "amount": {
                        "description": "Сумма для изменения баланса (может быть отрицательной)",
                        "type": "number",
                        "format": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
            "UserInfoResponse": {
                "title": "Ответ с информацией о пользователе",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Информация о пользователе получена успешно"
                    },
                    "data": {
                        "$ref": "#/components/schemas/UserWithBalance"
                    }
                },
                "type": "object"
            },
            "UserBalanceResponse": {
                "title": "Ответ с балансом пользователя",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Баланс пользователя получен успешно"
                    },
                    "data": {
                        "description": "Данные баланса пользователя",
                        "properties": {
                            "id": {
                                "description": "Идентификатор пользователя",
                                "type": "integer",
                                "example": 1
                            },
                            "balance": {
                                "description": "Баланс пользователя",
                                "type": "number",
                                "format": "integer",
                                "example": 1500
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Аутентификация через Laravel Sanctum токены",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}